fix: Use correct EntityNormalizer in D11
Apparently in Drupal 11 it is not enough to just set the supportedInterfaceOrClass property in the EntityNormalizers in order for the serialization API to find the correct Normalizer for a given Entity type.
Now we need to define the supported types in the getSupportedTypes method. The standard Drupal EntityNormalizer just returns the general EntityInterface::class here.
Since both the Normalizers for WisskiPathbuilders and WisskiEntities inherit from the standard EntityNormalizer they also declare EntityInterface::class as supported.
This would result in every Entity that is passed to the serialization API to be normalized by the WisskiEntityNormalizer, due to the high priority (1) set in the services.yml and the WisskiEntityNormalizer being in front of WisskiPathbuilderEntityNormalizer in alphanumerical odering.
Fix this by overwriting the getSupportedTypes method and make it return the respective classes instead of the general EntityInterface.